pyproject.toml
[project]
name = "rogier"
version = "0.1.0"
requires-python = ">=3.11,<3.13"
dependencies = [
    "fastapi>=0.110,<1.0",
    "uvicorn[standard]>=0.27,<1.0",
    "jinja2>=3.1,<4.0",
    "httpx>=0.26,<1.0",
    "itsdangerous>=2.1,<3.0",
    "bcrypt>=4.1,<5.0",
    "python-dotenv>=1.0,<2.0",
    "python-multipart>=0.0.9,<1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9.0",
    "pytest-asyncio>=0.23,<1.0",
    "httpx[cli]",
    "ruff>=0.3,<1.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "slow: tests lents nécessitant le CSA complet",
]
